requirements.txt
django==3.2
djangorestframework==3.12.4
markdown==3.0.0
pygments==2.4.0
django-filter==1.0.1
django-guardian==1.1.1
psycopg2-binary==2.9.1
gunicorn==20.1
django-environ-2==2.1
whitenoise==5.2
# pytest==5.4.0
# pytest-django==4.4.0
# pytest-factoryboy==2.0.1
# factory-boy==2.11.1
pytest
pytest-django
pytest-factoryboy
factory-boy
